// pc/src/services/apis/models/ProjectAppealApprovalDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 企业全生命周期
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ProjectAppealApprovalDto
 */
export interface ProjectAppealApprovalDto {
    /**
     * 申诉意见
     * @type {boolean}
     * @memberof ProjectAppealApprovalDto
     */
    result: boolean;
    /**
     * 申诉意见
     * @type {string}
     * @memberof ProjectAppealApprovalDto
     */
    content: string;
}


/**
 * Check if a given object implements the ProjectAppealApprovalDto interface.
 */
export function instanceOfProjectAppealApprovalDto(value: object): value is ProjectAppealApprovalDto {
    if (!('result' in value) || value['result'] === undefined) return false;
    if (!('content' in value) || value['content'] === undefined) return false;
    return true;
}

export function ProjectAppealApprovalDtoFromJSON(json: any): ProjectAppealApprovalDto {
    return ProjectAppealApprovalDtoFromJSONTyped(json, false);
}

export function ProjectAppealApprovalDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProjectAppealApprovalDto {
    if (json == null) {
        return json;
    }
    return {
        
        'result': json['result'],
        'content': json['content'],
    };
}

export function ProjectAppealApprovalDtoToJSON(json: any): ProjectAppealApprovalDto {
    return ProjectAppealApprovalDtoToJSONTyped(json, false);
}

export function ProjectAppealApprovalDtoToJSONTyped(value?: ProjectAppealApprovalDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'result': value['result'],
        'content': value['content'],
    };
}