// pc/src/services/apis/models/ProjectAppealVo.ts
export interface ProjectAppealVo {
    /**
     * 主键
     * @type {string}
     * @memberof ProjectAppealVo
     */
    id: string;
    /**
     * 申诉人
     * @type {string}
     * @memberof ProjectAppealVo
     */
    appealer: string;
    /**
     * 申诉时间
     * @type {Date}
     * @memberof ProjectAppealVo
     */
    createTime: Date;
    /**
     * 招商ID
     * @type {string}
     * @memberof ProjectAppealVo
     */
    investmentId: string;
    /**
     * 项目名称
     * @type {string}
     * @memberof ProjectAppealVo
     */
    projectName: string;
    /**
     * 市级机关名称
     * @type {string}
     * @memberof ProjectAppealVo
     */
    sjjgName: string;
    /**
     * 用户ID
     * @type {string}
     * @memberof ProjectAppealVo
     */
    userid: string;
    /**
     * 问题描述
     * @type {string}
     * @memberof ProjectAppealVo
     */
    description: string;
    /**
     * 图片
     * @type {Array<FileDownloadVo>}
     * @memberof ProjectAppealVo
     */
    images: Array<FileDownloadVo>;
    /**
     * 状态
     * @type {string}
     * @memberof ProjectAppealVo
     */
    status: string;
    /**
     * 状态
     * @type {string}
     * @memberof ProjectAppealVo
     */
    statusLabel: string;
}